tests: Added unit test for workspace use-cases

+++ api/src/application/use-cases/workspace/deleteWorkspaceUseCase.js
@@ -6,6 +6,11 @@ class DeleteWorkspaceUseCase {
   }
 
   async execute(workspaceId) {
+    if (!workspaceId) {
+      throw boom.badRequest(
+        'The ID of the workspace to be deleted was not provided.',
+      );
+    }
     const deletedWorkspace = await this.workspaceRepository.delete(workspaceId);
     if (deletedWorkspace === 0) throw boom.notFound('Workspace not found');
